Fix up .gitignore issues and make halogen run
It's possible that many previous commits are bad because
of .gitignore issues after I switched up my module structure.

As of this commit, we have halogen running from inside a react app,
which in the process demonstrated one way (probably not the best
way; it was difficult to figure out anything that worked) to run
Aff from javascript

This is probably my final commit of this simple demo (although I may
revisit it using other libraries, or adding CSS, etc) other than
README updates.